// FE/src/types/module/module.d.ts
import { tableOperationType } from "../opearation/operation";

export interface tableModuleGroupData {
  id?: string;
  moduleName: string;
  moduleCode: string;
  codeObject: string;
  operations?: tableOperationType[];
  selectedCodes?: undefined;
}

interface TableRowHeader {
  //   lstOperation: tableOperationType[]
  isAllChecked: boolean;
  isIndeterminate: boolean;
  moduleCode: string;
}